requirements.txt
blinker==1.9.0
click==8.3.1
dotenv==0.9.9
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.5
